+++ langs/dotnet/lang-{{cookiecutter.lang_name}}/validate.bash
@@ -6,19 +6,16 @@ set -o pipefail
 
 # shellcheck disable=SC1091
 source "$POLY_SRC/lib/lib-util.bash"
-if [[ -e "$POLYGLOT_ROOT/.tasks/task-util.bash" ]]; then
-    source "$POLYGLOT_ROOT/.tasks/task-util.bash"
-fi
+# shellcheck disable=SC1091
+[[ -e "$POLYGLOT_ROOT/.tasks/task-util.bash" ]] && source "$POLYGLOT_ROOT/.tasks/task-util.bash"
 
 tdot "dotnet" "Validating"
-if [[ -e "$POLYGLOT_ROOT/blah.sln" ]]; then
-    rm "$POLYGLOT_ROOT/blah.sln"
-fi
+[[ -e "$POLYGLOT_ROOT/blah.sln" ]] && rm "$POLYGLOT_ROOT/blah.sln"
 
-dotnet new sln                 || fail "Dotnet Failed"
-dotnet solution add src/cs_exe || fail "Dotnet Failed"
-dotnet solution add src/cs_lib || fail "Dotnet Failed"
-dotnet solution add src/fs_exe || fail "Dotnet Failed"
-dotnet build                   || fail "Dotnet Failed"
+dotnet new sln                 || fail "Dotnet new sln Failed"
+dotnet solution add src/cs_exe || fail "Dotnet add cs_exe Failed"
+dotnet solution add src/cs_lib || fail "Dotnet add cs_lib Failed"
+dotnet solution add src/fs_exe || fail "Dotnet add fs_exe Failed"
+dotnet build                   || fail "Dotnet build Failed"
 
 tdot "dotnet" "TODO validate docfx"
