+++ langs/python/lang-{{cookiecutter.lang_name}}/doc.bash
@@ -6,18 +6,12 @@ set -o pipefail
 
 # shellcheck disable=SC1091
 source "$POLY_SRC/lib/lib-util.bash"
-if [[ -e "$POLYGLOT_ROOT/.tasks/task-util.bash" ]]; then
-    # shellcheck disable=SC1091
-    source "$POLYGLOT_ROOT/.tasks/task-util.bash"
-fi
+# shellcheck disable=SC1091
+[[ -e "$POLYGLOT_ROOT/.tasks/task-util.bash" ]] && source "$POLYGLOT_ROOT/.tasks/task-util.bash"
 
 function check () {
-    if [[ ! -e "$POLYGLOT_SPHINX_CONF_DIR/conf.py" ]]; then
-        fail "NOT FOUND: $POLYGLOT_SPHINX_CONF_DIR/conf.py"
-    fi
-    if [[ -z "${POLYGLOT_DOCS:-}" ]]; then
-        fail "NOT DEFINED: POLYGLOT_DOCS"
-    fi
+    [[ -e "$POLYGLOT_SPHINX_CONF_DIR/conf.py" ]] || fail "NOT FOUND: $POLYGLOT_SPHINX_CONF_DIR/conf.py"
+    [[ -n "${POLYGLOT_DOCS:-}" ]] || fail "NOT DEFINED: POLYGLOT_DOCS"
 }
 
 subhead "[python]" "Building Sphinx"
@@ -31,10 +25,9 @@ echo -e "
 - builder          : ${SPHINX_BUILDER:-html}
 "
 
-if [[ -d "${SPHINX_OUT}" ]]; then
-    rm -r "${SPHINX_OUT}"
-fi
-uv run --frozen sphinx-build \
+[[ -d "${SPHINX_OUT}" ]] && rm -r "${SPHINX_OUT}"
+
+( uv run --frozen sphinx-build \
     --verbose \
     --write-all \
     --fresh-env \
@@ -44,4 +37,4 @@ uv run --frozen sphinx-build \
     --builder "${SPHINX_BUILDER:-html}" \
     "$SRC_DIR" \
     "$SPHINX_OUT"
-    # || fail "Sphinx Failed"
+  ) || fail "Sphinx Failed"
